chore(config): split config.py into a cohesive config/ package
Pure relocation of the 1444-LOC config.py into a config/ package (no documented
surface change — every teatree.config.<name> import and patch target resolves
unchanged via the __init__ facade): enums (the four config enums), settings
(UserSettings/TeaTreeConfig/OverlayEntry/E2ERepo + the _parse_* coercers + the
override registries), loader (load_config + the toml/logging/dir entry points),
discovery (overlay discovery), and resolution (get_effective_settings +
cadence_seconds + the autonomy collapse + the per-setting resolvers).

_resolve_autonomy / _resolve_speed collapse into one generic
_resolve_enum_setting(teatree, key, enum, default); the tuple-returning
_resolve_on_behalf_post_mode and the nested-table
_resolve_slack_voice_classifier_mode stay bespoke.

Deviation from the spec's 3-file target: 6 files, forced by the module-health
500-LOC cap AND the RUF067 init-is-re-exports-only rule (load_config can't live
in __init__). Submodules reach each other's non-private load_config / discover_* /
CONFIG_PATH through the package facade at call-time (breaks the cycle, preserves
patch semantics); the two private discovery helpers are called same-module and
their few test patches repoint to the teatree.config.discovery def-site.

config_speak stays top-level (foundation layer); not moved.

Campaign: architecture-perfect PR5 (god-module splits), config concern.
